Fix NativeAudioNodeDelay
Former-commit-id: 277f2d90139e8fb7cdbed4dd1a10b14896a7e3fc [formerly f94152a6355ebceb6c55d49387ce9708e64826f0]
Former-commit-id: 17aab26701722b0a94d05e988c5745d10f268570

+++ av/NativeAudioNode.h
@@ -208,8 +208,8 @@ class NativeAudioNodeDelay : public NativeAudioNode
         };
 
         double delay;
-        double wet;
-        double dry;
+        int wet;
+        int dry;
         float **buffers;
 
         virtual bool process();
